requirements.txt
Django==2.2.4
django-ckeditor==5.7.1
django-datetimepicker==3.14
django-js-asset==1.2.2
django-mptt==0.10.0
pytz==2019.2
sqlparse==0.3.0
